+++ lib/sp3d_renderer.dart
@@ -21,7 +21,7 @@ import 'package:simple_3d_renderer/sp3d_v2d.dart';
 ///
 class Sp3dRenderer extends StatefulWidget {
   final String className = 'Sp3dRenderer';
-  final String version = '4';
+  final String version = '5';
   final GlobalKey key;
   final Size size;
   final Sp3dV2D worldOrigin;
@@ -29,6 +29,17 @@ class Sp3dRenderer extends StatefulWidget {
   final Sp3dCamera camera;
   final Sp3dLight light;
   final bool useUserGesture;
+  final bool allowFullCtrl;
+  final bool allowUserWorldRotation;
+  final bool checkTouchObj;
+  final ValueNotifier<int>? vn;
+
+  // タッチリスナの定義。速度の問題があるので、onPanDownでのみ当たり判定の演算を行い、当たり判定があれば情報クラスを返す。
+  final Function(Offset, Sp3dFaceObj?)? onPanDownListener;
+  final Function()? onPanCancelListener;
+  final Function(Offset)? onPanStartListener;
+  final Function(Offset)? onPanUpdateListener;
+  final Function()? onPanEndListener;
 
   /// Constructor
   /// * [key] : Global Key.
@@ -39,63 +50,34 @@ class Sp3dRenderer extends StatefulWidget {
   /// * [world] : Objects to be drawn.
   /// * [camera] : camera.
   /// * [light] : light.
-  /// * [useUserGesture] : If true, apply GestureDetector.
+  /// * [useUserGesture] : If true, apply GestureDetector. If false, also disabled allowFullCtrl flag.
+  /// * [allowFullCtrl] : If true, apply full listener. If false, apply world rotation by user.
+  /// * [allowUserWorldRotation] : If true, allow world rotation by user when allowFullCtrl is true.
+  /// However, when checkTouchObj is true and touch target is exist when onPanDown, the subsequent rotation is suppressed.
+  /// * [checkTouchObj] : If this is true and allowFullCtrl is true, Returns the information of the object touched when onPanDown.
+  /// If false, and if there is no touch target, null is returned by onPanDown.
+  /// * [onPanDownListener] : Callback at user touch initiation. The arguments of the function to be set are the offset and the object touched at the time of touch.
+  /// * [onPanCancelListener] : Callback at user touch cancel. It is called instead of onTouchEnd when user touch cancel.
+  /// * [onPanStartListener] : Callback at user pan start.
+  /// * [onPanUpdateListener] : Callback at user pan update.
+  /// * [onPanEndListener] : Callback at user pan end.
+  /// * [vn] : ValueNotifier. If update this notifier.value, custom painter in renderer will repaint.
   Sp3dRenderer(this.key, this.size, this.worldOrigin, this.world, this.camera,
       this.light,
-      {this.useUserGesture = true})
+      {this.useUserGesture = true,
+      this.allowFullCtrl = false,
+      this.allowUserWorldRotation = true,
+      this.checkTouchObj = true,
+      this.onPanDownListener,
+      this.onPanCancelListener,
+      this.onPanStartListener,
+      this.onPanUpdateListener,
+      this.onPanEndListener,
+      this.vn})
       : super(key: key);
 
   @override
   _Sp3dRendererState createState() => _Sp3dRendererState();
-
-  /// (en)Deep copy the object.
-  ///
-  /// (ja)このオブジェクトをディープコピーします。
-  ///
-  /// * [key] : New GlobalKey.
-  Sp3dRenderer deepCopy(GlobalKey key) {
-    return Sp3dRenderer(key, this.size, this.worldOrigin.deepCopy(),
-        this.world.deepCopy(), this.camera.deepCopy(), this.light.deepCopy(),
-        useUserGesture: this.useUserGesture);
-  }
-
-  /// (en)Convert the object to a dictionary.
-  ///
-  /// (ja)このオブジェクトを辞書に変換します。
-  ///
-  Map<String, dynamic> toDict() {
-    Map<String, dynamic> d = {};
-    d['class_name'] = this.className;
-    d['version'] = this.version;
-    d['size'] = [this.size.width, this.size.height];
-    d['world_origin'] = this.worldOrigin.toDict();
-    d['world'] = this.world.toDict();
-    d['camera'] = this.camera.toDict();
-    d['lights'] = this.light.toDict();
-    d['use_user_gesture'] = this.useUserGesture;
-    return d;
-  }
-
-  /// (en)Restore this object from the dictionary.
-  ///
-  /// (ja)辞書からオブジェクトを復元します。
-  ///
-  /// * [key] : GlobalKey.
-  /// * [src] : A dictionary made with to_dict of this class.
-  static Sp3dRenderer fromDict(GlobalKey key, Map<String, dynamic> src) {
-    List<Sp3dObj> objs = [];
-    for (Map<String, dynamic> i in src['sp3d_objs']) {
-      objs.add(Sp3dObj.fromDict(i));
-    }
-    return Sp3dRenderer(
-        key,
-        Size(src['size'][0], src['size'][1]),
-        Sp3dV2D.fromDict(src['world_origin']),
-        Sp3dWorld.fromDict(src['world']),
-        Sp3dCamera.fromDict(src['camera']),
-        Sp3dLight.fromDict(src['light']),
-        useUserGesture: src['use_user_gesture']);
-  }
 }
 
 class _Sp3dRendererState extends State<Sp3dRenderer> {
@@ -108,46 +90,119 @@ class _Sp3dRendererState extends State<Sp3dRenderer> {
   // 現在の回転角
   double _angle = 0.0;
 
+  // 当たり判定計算用のPath
+  final Path _p = Path();
+
+  // 以降の回転を抑制するかどうかのフラグ
+  bool _canRotation = true;
+
   @override
   Widget build(BuildContext context) {
     if (widget.useUserGesture) {
-      return GestureDetector(
-        child: CustomPaint(
-          painter: _Sp3dCanvasPainter(widget),
-          size: widget.size,
-        ),
-        // ドラッグ操作開始
-        onPanStart: (DragStartDetails dsd) {
-          this._sp = Sp3dV3D(dsd.localPosition.dx, dsd.localPosition.dy, 0);
-          this._preP = Sp3dV3D(dsd.localPosition.dx, dsd.localPosition.dy, 0);
-        },
-        // ドラッグ操作中の変化
-        onPanUpdate: (DragUpdateDetails dud) {
-          setState(() {
-            final Sp3dV3D nowP =
-                Sp3dV3D(dud.localPosition.dx, dud.localPosition.dy, 0);
-            final Sp3dV3D moveP = nowP - this._preP;
-            this._angle += moveP.len();
-            if (this._angle > 360) {
-              this._angle %= 360;
+      if (widget.allowFullCtrl) {
+        return GestureDetector(
+          child: CustomPaint(
+            painter: _Sp3dCanvasPainter(widget),
+            size: widget.size,
+          ),
+          onPanDown: (DragDownDetails d) {
+            this._canRotation = true;
+            if (widget.onPanDownListener != null) {
+              if (widget.checkTouchObj) {
+                for (Sp3dFaceObj i in widget.world.sortedAllFaces.reversed) {
+                  this._p.reset();
+                  if (i.vertices2d.length == 3) {
+                    this._p.moveTo(i.vertices2d[0].x, i.vertices2d[0].y);
+                    this._p.lineTo(i.vertices2d[1].x, i.vertices2d[1].y);
+                    this._p.lineTo(i.vertices2d[2].x, i.vertices2d[2].y);
+                    this._p.close();
+                    if (this._p.contains(d.localPosition)) {
+                      widget.onPanDownListener!(d.localPosition, i);
+                      this._canRotation = false;
+                      return;
+                    }
+                  } else {
+                    this._p.moveTo(i.vertices2d[0].x, i.vertices2d[0].y);
+                    this._p.lineTo(i.vertices2d[1].x, i.vertices2d[1].y);
+                    this._p.lineTo(i.vertices2d[2].x, i.vertices2d[2].y);
+                    this._p.lineTo(i.vertices2d[3].x, i.vertices2d[3].y);
+                    this._p.close();
+                    if (this._p.contains(d.localPosition)) {
+                      widget.onPanDownListener!(d.localPosition, i);
+                      this._canRotation = false;
+                      return;
+                    }
+                  }
+                }
+                widget.onPanDownListener!(d.localPosition, null);
+              } else {
+                widget.onPanDownListener!(d.localPosition, null);
+              }
             }
-            final Sp3dV3D diff = nowP - this._sp;
-            final Sp3dV3D axis = Sp3dV3D(diff.y, diff.x, 0).nor();
-            widget.camera.rotate(axis, this._angle * pi / 180);
-            _preP = nowP;
-          });
-        },
-        // ドラッグ操作終了時
-        onPanEnd: (DragEndDetails ded) {
-          // 何もしない
-        },
-      );
+          },
+          onPanCancel: () {
+            if (widget.onPanCancelListener != null) {
+              widget.onPanCancelListener!();
+            }
+          },
+          onPanStart: (DragStartDetails d) {
+            if (widget.allowUserWorldRotation && this._canRotation) {
+              this._sp = Sp3dV3D(d.localPosition.dx, d.localPosition.dy, 0);
+              this._preP = Sp3dV3D(d.localPosition.dx, d.localPosition.dy, 0);
+            }
+            if (widget.onPanStartListener != null) {
+              widget.onPanStartListener!(d.localPosition);
+            }
+          },
+          onPanUpdate: (DragUpdateDetails d) {
+            if (widget.allowUserWorldRotation && this._canRotation) {
+              _rotation(d.localPosition);
+            }
+            if (widget.onPanUpdateListener != null) {
+              widget.onPanUpdateListener!(d.localPosition);
+            }
+          },
+          onPanEnd: (DragEndDetails d) {
+            if (widget.onPanEndListener != null) {
+              widget.onPanEndListener!();
+            }
+          },
+        );
+      } else {
+        return GestureDetector(
+            child: CustomPaint(
+              painter: _Sp3dCanvasPainter(widget),
+              size: widget.size,
+            ),
+            onPanStart: (DragStartDetails d) {
+              this._sp = Sp3dV3D(d.localPosition.dx, d.localPosition.dy, 0);
+              this._preP = Sp3dV3D(d.localPosition.dx, d.localPosition.dy, 0);
+            },
+            onPanUpdate: (DragUpdateDetails d) {
+              _rotation(d.localPosition);
+            });
+      }
     } else {
       return CustomPaint(
         painter: _Sp3dCanvasPainter(widget),
         size: widget.size,
       );
     }
+  }
+
+  /// world rotation
+  void _rotation(Offset offset) {
+    setState(() {
+      final Sp3dV3D nowP = Sp3dV3D(offset.dx, offset.dy, 0);
+      this._angle += (nowP - this._preP).len();
+      if (this._angle > 360) {
+        this._angle %= 360;
+      }
+      final Sp3dV3D diff = nowP - this._sp;
+      final Sp3dV3D axis = Sp3dV3D(diff.y, diff.x, 0).nor();
+      this.widget.camera.rotate(axis, this._angle * pi / 180);
+      this._preP = nowP;
+    });
   }
 }
 
@@ -156,7 +211,7 @@ class _Sp3dCanvasPainter extends CustomPainter {
   final Paint p = Paint();
   final Path path = Path();
 
-  _Sp3dCanvasPainter(this.w);
+  _Sp3dCanvasPainter(this.w) : super(repaint: w.vn);
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -166,6 +221,9 @@ class _Sp3dCanvasPainter extends CustomPainter {
         this.w.camera.getPrams(this.w.world, this.w.worldOrigin);
     // z軸を基準にして遠いところから順番に塗りつぶすために全てのfaceを逆順ソート。
     allFaces.sort((Sp3dFaceObj a, Sp3dFaceObj b) => b.dist.compareTo(a.dist));
+    if (w.allowFullCtrl) {
+      this.w.world.sortedAllFaces = allFaces;
+    }
     // 描画
     for (Sp3dFaceObj fo in allFaces) {
       // パスを描画